Rename to avoid conflicts with similarly named local implementations
