+++ osu.Framework.Tests/IO/TestLogging.cs
@@ -6,7 +6,7 @@ using System.Threading;
 using System.Threading.Tasks;
 using NUnit.Framework;
 using osu.Framework.Logging;
-using osu.Framework.Platform;
+using osu.Framework.Testing;
 
 namespace osu.Framework.Tests.IO
 {
@@ -53,7 +53,7 @@ namespace osu.Framework.Tests.IO
 
             try
             {
-                using (var host = new TestHeadlessGameHost())
+                using (var host = new TestRunHeadlessGameHost())
                 {
                     var game = new TestGame();
                     game.Schedule(() => throw new TestException());
@@ -88,7 +88,7 @@ namespace osu.Framework.Tests.IO
         /// <param name="fireCount">How many exceptions to fire.</param>
         private void runWithIgnoreCount(int ignoreCount, int fireCount)
         {
-            using (var host = new TestHeadlessGameHost())
+            using (var host = new TestRunHeadlessGameHost())
             {
                 host.ExceptionThrown += ex => ignoreCount-- > 0;
 
@@ -107,7 +107,7 @@ namespace osu.Framework.Tests.IO
         {
             Assert.Throws<TestException>(() =>
             {
-                using (var host = new TestHeadlessGameHost())
+                using (var host = new TestRunHeadlessGameHost())
                     host.Run(new CrashTestGame());
             });
         }
@@ -124,7 +124,7 @@ namespace osu.Framework.Tests.IO
         [Test]
         public void TestGameUnobservedExceptionDoesntCrashGame()
         {
-            using (var host = new TestHeadlessGameHost())
+            using (var host = new TestRunHeadlessGameHost())
             {
                 TaskCrashTestGame game = new TaskCrashTestGame();
                 host.Run(game);
