+++ osu.Framework.Tests/Platform/GameHostSuspendTest.cs
@@ -9,6 +9,7 @@ using osu.Framework.Bindables;
 using osu.Framework.Configuration;
 using osu.Framework.Development;
 using osu.Framework.Platform;
+using osu.Framework.Testing;
 using osu.Framework.Threading;
 
 namespace osu.Framework.Tests.Platform
@@ -81,7 +82,7 @@ namespace osu.Framework.Tests.Platform
             Assert.AreEqual(GameThreadState.Exited, updateThreadState.Value);
         }
 
-        private class ExecutionModeGameHost : TestHeadlessGameHost
+        private class ExecutionModeGameHost : TestRunHeadlessGameHost
         {
             private readonly ExecutionMode threadMode;
