+++ osu.Framework/Testing/TestScene.cs
@@ -42,7 +42,7 @@ namespace osu.Framework.Testing
         [OneTimeSetUp]
         public void SetupGameHost()
         {
-            host = new TestHeadlessGameHost($"{GetType().Name}-{Guid.NewGuid()}");
+            host = new TestRunHeadlessGameHost($"{GetType().Name}-{Guid.NewGuid()}");
             runner = CreateRunner();
 
             if (!(runner is Game game))
